cleanup

+++ fatigue/apps/hrm_app.h
@@ -114,7 +114,7 @@ uint16_t hrm_get_resp_rate(void);
 /**************************************************************************//**
  * @brief This function returns the current respiratory intensity.
  *****************************************************************************/
-uint16_t hrm_get_resp_intensity(void);
+uint8_t hrm_get_resp_intensity(void);
 
 /**************************************************************************//**
  * @brief This function returns the Respiratory Fatigue indication status
